refactor to make tests runnable in node.js
