+++ fs.js
@@ -1,5 +1,1 @@
-/**
- * @notice This is a rexports of `fs/index.js` so consumers will
- * need to only `import * as fs from '@socketsupply/io/fs.js'`
- */
-export * from './fs/index.js'
+export * from 'fs'
