+++ index.js
@@ -12,6 +12,6 @@ export * as ipc from './ipc.js'
 export * as net from './net.js'
 export * as os from './os.js'
 export * as crypto from './crypto.js'
-export { default as process } from './process.js'
+export * as process from './process.js'
 export * as stream from './stream.js'
 export * as util from './util.js'
