+++ sdk/fs.js
@@ -1,0 +1,6 @@
+/**
+ * @notice This is a rexports of `fs/index.js` so consumers will
+ * need to only `import * as fs from '@socketsupply/io/fs.js'`
+ */
+
+export * from './fs/index.js'
