+++ sdk/fs/stream.js
@@ -1,4 +1,4 @@
-import { Readable, Writable } from '../stream.js'
+import { Readable, Writable } from '../../stream.js'
 import { Buffer } from 'buffer'
 
 export const DEFAULT_STREAM_HIGH_WATER_MARK = 16 * 1024
